set up global Modal, refactored filterModal to use global Modal, sorry :(
